# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  stats.CouponsBySizeResponse:
    properties:
      size_21x30:
        type: integer
      size_30x40:
        type: integer
      size_40x40:
        type: integer
      size_40x50:
        type: integer
      size_40x60:
        type: integer
      size_50x70:
        type: integer
    type: object
  stats.CouponsByStatusResponse:
    properties:
      activated:
        type: integer
      completed:
        type: integer
      new:
        type: integer
      used:
        type: integer
    type: object
  stats.CouponsByStyleResponse:
    properties:
      grayscale:
        type: integer
      max_colors:
        type: integer
      pop_art:
        type: integer
      skin_tones:
        type: integer
    type: object
  stats.GeneralStatsResponse:
    properties:
      activation_rate:
        type: number
      active_partners_count:
        type: integer
      last_updated:
        type: string
      total_coupons_activated:
        type: integer
      total_coupons_created:
        type: integer
      total_partners_count:
        type: integer
    type: object
...
  stats.TopPartnersResponse:
    properties:
      partners:
        items:
          $ref: '#/definitions/stats.TopPartnerItem'
        type: array
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API для системы мозаичных купонов
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Mosaic API
  version: "1.0"
paths:
...
  /admin/stats/coupons-by-size:
    get:
      description: Returns count of coupons by their sizes
      parameters:
      - description: Partner ID (optional)
        format: uuid
        in: query
        name: partner_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Coupons by size statistics
          schema:
            $ref: '#/definitions/stats.CouponsBySizeResponse'
        "400":
          description: Invalid partner ID format
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized access
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error while getting coupons by size statistics
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get coupons statistics by size
      tags:
      - admin-stats
  /admin/stats/coupons-by-status:
    get:
      description: Returns count of coupons in each status
      parameters:
      - description: Partner ID (optional)
        format: uuid
        in: query
        name: partner_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Coupons by status statistics
          schema:
            $ref: '#/definitions/stats.CouponsByStatusResponse'
        "400":
          description: Invalid partner ID format
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized access
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error while getting coupons by status statistics
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get coupons statistics by status
      tags:
      - admin-stats
  /admin/stats/coupons-by-style:
    get:
      description: Returns count of coupons by their processing styles
      parameters:
      - description: Partner ID (optional)
        format: uuid
        in: query
        name: partner_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Coupons by style statistics
          schema:
            $ref: '#/definitions/stats.CouponsByStyleResponse'
        "400":
          description: Invalid partner ID format
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized access
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error while getting coupons by style statistics
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get coupons statistics by style
      tags:
      - admin-stats
  /admin/stats/general:
    get:
      description: Returns general system statistics including coupon count, partner
        count, and activation percentage
      produces:
      - application/json
      responses:
        "200":
          description: General statistics data
          schema:
            $ref: '#/definitions/stats.GeneralStatsResponse'
        "401":
          description: Unauthorized access
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error while getting general statistics
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get general statistics
      tags:
      - admin-stats
  /admin/stats/partners:
    get:
      description: Returns statistics for all partners in the system
      produces:
      - application/json
      responses:
        "200":
          description: Partners statistics data
          schema:
            $ref: '#/definitions/stats.PartnerListStatsResponse'
        "401":
          description: Unauthorized access
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error while getting all partners statistics
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get all partners statistics
      tags:
      - admin-stats
  /admin/stats/partners/{partner_id}:
    get:
      description: Returns detailed statistics for a specific partner
      parameters:
      - description: Partner ID
        format: uuid
        in: path
        name: partner_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Partner statistics data
          schema:
            $ref: '#/definitions/stats.PartnerStatsResponse'
        "400":
          description: Invalid partner ID format
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized access
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Partner not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error while getting partner statistics
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get partner statistics
      tags:
      - admin-stats
  /admin/stats/realtime:
    get:
      consumes:
      - application/json
      description: Provides real-time statistics updates via WebSocket connection
      produces:
      - application/json
      responses:
        "101":
          description: Switching to WebSocket protocol
          schema:
            type: string
        "400":
          description: WebSocket upgrade failed
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized access
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get real-time statistics via WebSocket
      tags:
      - admin-stats
...
  /admin/stats/time-series:
    get:
      description: Returns data for building time-based charts and graphs
      parameters:
      - description: Partner ID (optional)
        format: uuid
        in: query
        name: partner_id
        type: string
      - description: Start date (YYYY-MM-DD)
        format: date
        in: query
        name: date_from
        type: string
      - description: End date (YYYY-MM-DD)
        format: date
        in: query
        name: date_to
        type: string
      - description: Grouping period
        enum:
        - day
        - week
        - month
        - year
        in: query
        name: period
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Time series statistics data
          schema:
            $ref: '#/definitions/stats.TimeSeriesStatsResponse'
        "400":
          description: Invalid request parameters format
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized access
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error while getting time series statistics
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get time series statistics
      tags:
      - admin-stats
  /admin/stats/top-partners:
    get:
      description: Returns list of top partners by activity metrics
      parameters:
      - description: 'Number of partners to return (default: 10)'
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Top partners statistics
          schema:
            $ref: '#/definitions/stats.TopPartnersResponse'
        "400":
          description: Invalid limit parameter
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized access
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error while getting top partners
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get top partners by activity
      tags:
      - admin-stats
...
  /partner/stats/my:
    get:
      description: Returns statistics for the current authenticated partner
      produces:
      - application/json
      responses:
        "200":
          description: Partner statistics data
          schema:
            $ref: '#/definitions/stats.PartnerStatsResponse'
        "401":
          description: Partner not authenticated
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error while getting partner statistics
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get my partner statistics
      tags:
      - partner-stats
  /partner/stats/my/coupons-by-status:
    get:
      description: Returns coupons statistics by status for the current authenticated
        partner
      produces:
      - application/json
      responses:
        "200":
          description: Partner coupons by status statistics
          schema:
            $ref: '#/definitions/stats.CouponsByStatusResponse'
        "401":
          description: Partner not authenticated
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error while getting coupons by status statistics
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get my partner coupons statistics by status
      tags:
      - partner-stats
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"